LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/module-tla-promise-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected keyword 'new'

PASS document.write in an imported module
